+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/DocumentModelAdministrationClient/DocumentClassifierAdministrationLiveTests.cs
@@ -68,7 +68,7 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
             }
             finally
             {
-                if (operation.HasValue)
+                if (operation != null && operation.HasValue)
                 {
                     await client.DeleteDocumentClassifierAsync(classifierId);
                 }
@@ -113,7 +113,7 @@ namespace Azure.AI.FormRecognizer.DocumentAnalysis.Tests
             }
             finally
             {
-                if (operation.HasValue)
+                if (operation != null && operation.HasValue)
                 {
                     await client.DeleteDocumentClassifierAsync(classifierId);
                 }
